+++ tests/components/test_header.py
@@ -228,7 +228,7 @@ def test_header_with_service_name_and_navigation(env):
 
 
 def test_header_with_service_name_and_navigation_and_autoescape(env, loader):
-    from govuk_frontend.templates import Environment
+    from govuk_frontend_jinja.templates import Environment
     env = Environment(autoescape=True, loader=loader)
     template = env.from_string(
 """
